+++ packages/plugins/other/schema-ast/tests/schema-ast.spec.ts
@@ -8,6 +8,44 @@ import { codegen } from '@graphql-codegen/core';
 const SHOULD_THROW_ERROR = 'SHOULD_THROW_ERROR';
 
 describe('Schema AST', () => {
+  describe('Issues', () => {
+    it('#4919 - should support sorting the schema', async () => {
+      const schema = buildSchema(/* GraphQL */ `
+        type Query {
+          d: String
+          z: String
+          a: String
+        }
+
+        type User {
+          aa: String
+          a: String
+        }
+
+        type A {
+          s: String
+          b: String
+        }
+      `);
+      const content = await plugin(schema, [], { sort: true });
+      expect(content).toBeSimilarStringTo(`
+      type A {
+        b: String
+        s: String
+      }
+      
+      type Query {
+        a: String
+        d: String
+        z: String
+      }
+      
+      type User {
+        a: String
+        aa: String
+      }`);
+    });
+  });
   describe('Validation', () => {
     it('Should enforce graphql extension when its the only plugin', async () => {
       const fileName = 'output.ts';