+++ packages/plugins/typescript/typed-document-node/tests/typed-document-node.spec.ts
@@ -1,0 +1,10 @@
+import { Types } from '@graphql-codegen/plugin-helpers';
+import { plugin } from '../src';
+
+describe('TypeDocumentNode', () => {
+  it('Should not output imports when there are no operations at all', async () => {
+    const result = (await plugin(null as any, [], {})) as Types.ComplexPluginOutput;
+    expect(result.content).toBe('');
+    expect(result.prepend.length).toBe(0);
+  });
+});